requirements.txt
aiofiles==23.2.1
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
alembic==1.13.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
bcrypt==4.1.2
beautifulsoup4==4.12.3
blinker==1.7.0
boto3==1.34.59
botocore==1.34.59
bs4==0.0.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
configparser==6.0.0
confuse==2.0.1
Deprecated==1.2.14
dill==0.3.8
distro==1.9.0
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
exceptiongroup==1.2.0
fastapi==0.115.2
Flask==3.0.1
Flask-Migrate==4.0.5
Flask-SQLAlchemy==3.1.1
frozenlist==1.4.1
gradescope-tool==0.1.2
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.6
importlib-metadata==7.0.1
importlib_resources==6.4.5
itsdangerous==2.1.2
Jinja2==3.1.3
jiter==0.6.1
jmespath==1.0.1
limits==3.13.0
Mako==1.3.1
MarkupSafe==2.1.4
multidict==6.0.5
mysql-connector-python==8.3.0
numpy==2.1.2
openai==1.52.0
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pdf2image==1.17.0
pillow==11.0.0
psycopg2-binary==2.9.9
py2flowchart==0.0.3
pyasn1==0.5.1
pydantic==2.9.2
pydantic-settings==2.6.0
pydantic_core==2.23.4
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-http-client==3.3.7
python-jose==3.3.0
python-multipart==0.0.5
pytz==2024.1
pywsse==0.1.5.2
PyYAML==6.0.2
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
s3transfer==0.10.0
sendgrid==6.11.0
six==1.16.0
slowapi==0.1.9
sniffio==1.3.0
soupsieve==2.6
SQLAlchemy==2.0.25
starkbank-ecdsa==2.2.0
starlette==0.40.0
tqdm==4.66.5
twilio==8.13.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.18
uvicorn==0.15.0
Werkzeug==3.0.1
wrapt==1.16.0
yarl==1.9.4
zipp==3.17.0